requirements.txt
Flask>=2.2.3
Pillow>=9.5.0
setuptools>=67.6.1
pystray>=0.19.4
pytest>=7.3.1
easycython>=1.0.7
